requirements.txt
Flask==2.0.1
pdfkit==1.0.0
python-dotenv==0.19.0
